pc5: fix retransmissions
sometimes, xfer only gets part of the packet. Althoug outstanding data was
flushed wit tcflush, the original data was read, afterwards.

instead of a dumb sleep we're now trying to read anything outstanding ...
runing into this read's timeout.

In turn to find this, the fixup debug message was enhanced, as well.